# Conflicts:
#	core/EE_Config.core.php
#	core/EE_Load_Espresso_Core.core.php
#	core/EE_Registry.core.php
#	core/EE_System.core.php

[touch:9795]

+++ core/services/database/TableManager.php
@@ -73,7 +73,7 @@ class TableManager extends \EE_Base
         $full_table_name = $this->getTableAnalysis()->ensureTableNameHasPrefix($table_name);
         $columns = $this->getTableColumns($table_name);
         if ( ! in_array($column_name, $columns)) {
-            $alter_query = "ALTER TABLE $full_table_name ADD $column_name $column_info";
+            $alter_query = "ALTER TABLE {$full_table_name} ADD {$column_name} {$column_info}";
             return $wpdb->query($alter_query);
         }
         return true;
@@ -93,16 +93,16 @@ class TableManager extends \EE_Base
     {
         global $wpdb;
         $table_name = $this->getTableAnalysis()->ensureTableNameHasPrefix($table_name);
-        $fieldArray = array();
+        $field_array = array();
         if ( ! empty($table_name)) {
-            $columns = $wpdb->get_results("SHOW COLUMNS FROM $table_name ");
+            $columns = $wpdb->get_results("SHOW COLUMNS FROM {$table_name} ");
             if ($columns !== false) {
                 foreach ($columns as $column) {
-                    $fieldArray[] = $column->Field;
+                    $field_array[] = $column->Field;
                 }
             }
         }
-        return $fieldArray;
+        return $field_array;
     }
 
 
@@ -120,7 +120,7 @@ class TableManager extends \EE_Base
         global $wpdb;
         if ($this->getTableAnalysis()->tableExists($table_name)) {
             $table_name = $this->getTableAnalysis()->ensureTableNameHasPrefix($table_name);
-            return $wpdb->query("DROP TABLE IF EXISTS $table_name");
+            return $wpdb->query("DROP TABLE IF EXISTS {$table_name}");
         }
         return 0;
     }
@@ -145,8 +145,12 @@ class TableManager extends \EE_Base
                 $tables_to_delete[] = $table_name;
             }
         }
-        global $wpdb;
-        $wpdb->query('DROP TABLE ' . implode(', ', $tables_to_delete));
+        if( ! empty( $tables_to_delete ) ) {
+            global $wpdb;
+            //make sure we only have a unique strings in the array.
+            $tables_to_delete = array_unique($tables_to_delete);
+            $wpdb->query('DROP TABLE ' . implode(', ', $tables_to_delete));
+        }
         return $tables_to_delete;
     }
 
@@ -155,26 +159,27 @@ class TableManager extends \EE_Base
     /**
      * Drops the specified index from the specified table. $table_name can
      * optionally start with $wpdb->prefix or not
+
      *
-     * @global \wpdb $wpdb
+*@global \wpdb       $wpdb
      * @param string $table_name
-     * @param string $indexName
-     * @return int
+     * @param string $index_name
+     * @return int the number of indexes dropped. False if there was a datbase error
      */
-    public function dropIndex($table_name, $indexName)
+    public function dropIndex($table_name, $index_name)
     {
         if (apply_filters('FHEE__EEH_Activation__drop_index__short_circuit', false)) {
-            return false;
+            return 0;
         }
         global $wpdb;
         $table_name = $this->getTableAnalysis()->ensureTableNameHasPrefix($table_name);
-        $index_exists_query = "SHOW INDEX FROM $table_name WHERE Key_name = '$indexName'";
+        $index_exists_query = "SHOW INDEX FROM {$table_name} WHERE key_name = '{$index_name}'";
         if (
             $this->getTableAnalysis()->tableExists($table_name)
             && $wpdb->get_var($index_exists_query)
                === $table_name //using get_var with the $index_exists_query returns the table's name
         ) {
-            return $wpdb->query("ALTER TABLE $table_name DROP INDEX $indexName");
+            return $wpdb->query("ALTER TABLE {$table_name} DROP INDEX {$index_name}");
         }
         return 0;
     }
@@ -184,21 +189,23 @@ class TableManager extends \EE_Base
     /**
      * Just creates the requested table. $table_name can
      * optionally start with $wpdb->prefix or not
+
      *
-     * @param string $table_name
-     * @param string $createSql defining the table's columns and indexes
-     * @param string $engine    (no need to specify "ENGINE=", that's implied)
+*@param string       $table_name
+     * @param string $create_sql defining the table's columns and indexes
+     * @param string $engine     (no need to specify "ENGINE=", that's implied)
      * @return void
      * @throws \EE_Error
      */
-    public function createTable($table_name, $createSql, $engine = 'MyISAM')
+    public function createTable($table_name, $create_sql, $engine = 'MyISAM')
     {
         // does $sql contain valid column information? ( LPT: https://regex101.com/ is great for working out regex patterns )
-        if (preg_match('((((.*?))(,\s))+)', $createSql, $valid_column_data)) {
+        if (preg_match('((((.*?))(,\s))+)', $create_sql, $valid_column_data)) {
             $table_name = $this->getTableAnalysis()->ensureTableNameHasPrefix($table_name);
-            $SQL = "CREATE TABLE $table_name ( $createSql ) ENGINE=$engine DEFAULT CHARSET=utf8 COLLATE=utf8_general_ci;";
             /** @var \wpdb $wpdb */
             global $wpdb;
+            $SQL = "CREATE TABLE {$table_name} ( {$create_sql} ) ENGINE={$engine} " . $wpdb->get_charset_collate();
+
             //get $wpdb to echo errors, but buffer them. This way at least WE know an error
             //happened. And then we can choose to tell the end user
             $old_show_errors_policy = $wpdb->show_errors(true);
@@ -218,10 +225,46 @@ class TableManager extends \EE_Base
                     __('The following table creation SQL does not contain valid information about the table columns: %1$s %2$s',
                         'event_espresso'),
                     '<br />',
-                    $createSql
+                    $create_sql
                 )
             );
         }
+    }
+
+
+
+    /**
+     * Drops the specified index if it's size differs from $desired_index_size.
+     * WordPress' dbdelta method doesn't automatically change index sizes, so this
+     * method can be used to only drop the index if needed, and afterwards dbdelta can be used as normal.
+     * If the table doesn't exist, or it exists but the index does not, or returns false
+     *
+     * @param string $table_name
+     * @param string $index_name
+     * @param string $column_name if none is provided, we assume the column name matches the index (often true in EE)
+     * @param string|int $desired_index_size defaults to TableAnalysis::index_col_size, the max for utf8mb4.
+     * @return bool whether an index was dropped or not
+     * @throws /EE_Error if table analysis object isn't defined
+     */
+    public function dropIndexIfSizeNot($table_name, $index_name, $column_name = null, $desired_index_size = TableAnalysis::INDEX_COLUMN_SIZE)
+    {
+        if($column_name === null){
+            $column_name = $index_name;
+        }
+        if(!$this->getTableAnalysis()->tableExists($table_name)){
+            return false;
+        }
+        $index_entries = $this->getTableAnalysis()->showIndexes($table_name,$index_name);
+        if(empty($index_entries)){
+            return false;
+        }
+        foreach($index_entries as $index_entry){
+            if( $column_name === $index_entry->Column_name
+                && (string)$desired_index_size !== $index_entry->Sub_part){
+                return $this->dropIndex($table_name,$index_name);
+            }
+        }
+        return false;
     }
 
 }
